Reorganize parser modules by purpose with backward-compat shims
- New `parser_tasks.py` holds task/paradigm-specific parsers: vividness
  (VVIQ-style) and the reality-monitoring (RM) word-pair task family,
  organized into clearly-labeled sections (encoding phase, single-chat
  full responses, trial-chain field components, trial-chain response
  classes, trial-chain unions, multi-phase task chain Task_1/2/3).
- New `parser_general.py` holds paradigm-agnostic parsers: generic
  Likert / agreement, generic response+rating shapes, word
  classification, task readiness, and the multi-phase TaskResponse
  Union (which references Task_1/2/3 from parser_tasks via a deferred
  forward reference + model_rebuild()).
- `parser.py` and `parser_extra.py` retained as backward-compat
  re-export shims: every name previously importable from those paths
  remains accessible there with no behavior change.
- `psychscanner.parsers` registry now scans the new themed modules.
  Class identity is preserved across all import paths (old shim, new
  themed module, top-level namespace).
- No class definitions, field names, or descriptions were modified —
  only file layout, ordering, and docstrings.
